Fix reading from closed shards (#65)
*Issue:*
https://github.com/awslabs/spark-sql-kinesis-connector/issues/64

*Description of changes:*
The Kinesis connector handles closed shards differently from open
shards. When the ShardSyncer encounters a closed shard, it adds it to
the shard map with a ShardEnd iterator type, rather than using the
initialPosition like it does for open shards.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.